monitor: use app options parsing helper code
This app is pretty ancient, so it was never converted to use the
option parsing helper code.  I'd like to add an option to this app
that takes an argument, and that's a pain to do when not using this
helper, so start by doing this conversion.

Review: https://reviewboard.asterisk.org/r/3429/
........

Merged revisions 412102 from file:///srv/subversion/repos/asterisk/trunk