res_stasis: Fix crash when handling a failed blind transfer message.
This changes fixes a crash that occurs when stasis determines if it
should send a message out to an application or not. The code
incorrectly assumed that a bridge snapshot would always be present
when in reality for failure cases it may not be.

ASTERISK-23573 #close
........

Merged revisions 412882 from http://svn.asterisk.org/svn/asterisk/branches/12