accountcode propagation: Initial patch that is on reviewboard r3601.
* Changed the accountcode propagation from simply setting the accountcode
of the channel initiating the call onto the outgoing channel to setting
the accountcode only if the outgoing channel doesn't already have one.
In other words, if a channel already has an accountcode it will not be
changed unless explicitly set by CHANNEL(accountcode) or an originate
method that can specify an accountcode.  The most visible effect the
change has is if the channel driver is configured with an accountcode
for new channels.

* Moved most accountcode propagation behavior into ast_request() which
eliminates several inconsistencies throughout the code.  (app_dial,
app_followme, app_queue, and the dialing API)

* Fixed the basic bridge sub-class updating peeraccount codes when the
number of channels in the bridge drops back down to two parties.

* Refactored ast_bridge_channel_update_accountcodes() to handle channels
joining/leaving the bridge.

* Fixed the basic bridge to not call the base pull method twice.

* Fixed CEL extracting the wrong ie value for the peeraccount.

* Fixed some incorrect CEL event ie types.  These seem to only be used
by the unit tests.