Implement a scheme for detecting loops/duplicated resources.
Each resource that is visited is added to a vector. When visiting
a new resource, we check if the resource exists in the vector.

It's not a very efficient algorithm since seeing if the resource
has been visited is an O(N) operation. It could be sped up by
using a different container type, though I'd first like to
verify the algorithm's correctness before optimizing.