Add requested debug messages from /r/3699.
This adds debug messages when building a resource tree so that
it can be seen how it is being built and where problems may
lie.

This also adds a debug message to an existing function that
indicates when a real subscription is being removed from the
master list of subscriptions.