Change algorithm for NOTIFY body generation.
Prior to this change, a "bottom-up" algorithm
was used. When a leaf subscription's state changed,
the body was generated and stored on the leaf subscription.
Then, moving up the tree, each parent list subscription
had its body regenerated until the root was reached. When
it came time to send a NOTIFY, the pre-generated body on
the root subscription was used in the NOTIFY body.

This would be a fine method to use if we were always
sending full or always sending partial notification
state. However, the context of the NOTIFY we send
can change whether we send a full or partial state.
Therefore, we have to generate the list body at the
time that we send our NOTIFY so we can generate the
proper body type.

The next large effort is to properly track the subscription
state of individual leaves/lists since right now that
is totally messed up.