Refactor the ast_sip_subscription into two separate objects.
I was going through an exercise where I was attempting to separate
the "real" and "virtual" parts of ast_sip_subscriptions more and
I realized that with only an exception or two, everything in the
structure was mutually exclusive to being in either the
real union member or virtual union member. I had a mini
"god object" on my hands. I have therefore separated the
ast_sip_subscription struct into two separate structures

ast_sip_subscription: This is the opaque forward-declared structure
in res_pjsip_pubsub that subscription handlers deal with. These are
essentially "virtual" subscriptions in that they have resource-specific
and handler-specific data on them only. No details about the actual
subscription are present.

sip_subscription_tree: This is the base on which ast_sip_subscriptions
stem. This contains information about the overarching subscription.
Within res_pjsip_pubsub, this structure is used much more often than
ast_sip_subscription.

This change is a breath of fresh air, because it eliminates a lot of
the conditional checking for whether the subscription was real or
virtual. Now, since they are separate types, the usage is much less
brittle than previous.

There are still checks in place to see if ast_sip_subscriptions are
lists or leaves. I'd like to eliminate that if possible, but I'm
not sure if I'll be able to do that.