media format attributes: Clean up modules, make pass through work in chan_sip
Josh did the bulk of this work on r3703 (thanks!). This patch includes that
review, as well as the work done in chan_sip in r3722 to make attributes
pass through. This included some additional changes to format_cap.

The major changes to format_cap are:
* format_cap now has a method to get the preferred codec out of a capabilities
  structure by some type. This prevents the situation where the
  'preferred codec' is not an audio codec (which can happen when format
   capabilities structures are copied, moved around, and perturbed).
* format_cap now explicitly prevents duplicate formats from being appended. A
  format is considered a duplicate if it has the same codec ID. This is a
  'strict' interpretation of equivalence, as it means that a format that
  contains an attribute cannot coexist with another format of the same type.
  This was needed, however, to prevent capabilities structures from having
  duplicate video formats (some with attributes, some without), which the
  channel drivers cannot yet understand or process (and which is not strictly
  needed right now).

The chan_sip channel driver has been tweaked up a bit to make sure that when a
capability is built out from an SDP, that capability (which is the jointcaps
structure) is given preference when constructing an outgoing SDP. That helps
ensure that adding a format to an outbound SDP gets the appropriate format w/
attributes, as opposed to the peer formats which would not have any attributes.

Note that since the jointcaps capability is guaranteed to be a subset of the
peer's capabilities, this ends up being generally the same.

Review: https://reviewboard.asterisk.org/r/3722/
Review: https://reviewboard.asterisk.org/r/3703/

ASTERISK-23957 #close