media_formats: Prevent crash during masquerade
When a masquerade occurs, a newly created channel replaces an existing channel
and steals its private data structure. This includes swapping the formats and
capabilities.

Since the newly created channel only contains ast_format_none, this causes an
assert to fire when the masqueraded channel is destroyed.

Removing the assert isn't a good idea. However, there's also no real need to
do the accessing that fires the asserts either: the masqueraded channel will
be destroyed, the references will be cleaned up appropriately, and life will go
on. free_translation is also called in a channel destructor, and again, things
will be cleaned up appropriately without going through the accessors that have
the asserts.

https://reviewboard.asterisk.org/r/3753/