Stasis: Allow prestart actions to be queued
This allows commands to be queued in a channel datastore to be executed
upon a channel entering Stasis(). This functionality is only available
from components of res_stasis.so. This functionality is required to get
a redirected channel back into the bridge for which it was destined due
to the attended transfer.

ASTERISK-23941
Review: https://reviewboard.asterisk.org/r/3729/